Publishing: Improve diff comparison when on custom branch (#1058)
Fix the publishing logic, so it works as expected in different scenarios. When doing prerelease in custom branch, it now compares whole content of the branch with the main/master branch to determine if a change has happened on a module. Old logic required a change to be on the last commit.