Removed reference
WhatIf deploymens of microsoft.sql/servers/ does not show output for databases if "server.location" is used. Changing it to "location" (which server.location is built from) solves the issue.

That being said it could be hard-coded to location: location instead of "location: contains(database, 'location') ? database.location : location" as it is not possible to deploy a database in a different location than the server, so it is a confusing parameter to have.